Fix updater stats task to stay until disk write is complete
The the updater progress stats is populated from
the docloader when documents are loaded into MapQueue.
Even after all documents are loaded to MapQueue, it takes
additional amount of time to sort records and and build/update
ondisk indexes. Even though loading is complete, we need to make
sure that the progress stats task should not dissapear until
the whole updater processing is complete.

Change-Id: Idf3d3304d488a0bf3851db1b2b172ea6e796f1e0