MB-11579 Use UPR_ADD_STREAM_FLAG_LATEST flag for upr stream request
When we request a stream for the first time and rollback happens due
to vbucket resets, there is race condition that could result in a
stream request with endseqno > vb_high_seqno of that vbucket. This
will cause stream to hang until more mutations are added to that vbucket.

Eg:
1. Requested stream for vb 228 for 0 - 6 and we received data
2. Requested stream for vb 229, but server told to rollback and we rolled
   back our entire index (Now, vb 228 sequence in index = 0, vbuuid =0 because
   of rollback)
3. To start index update, we obtained end_seqno from vbucket-seqno stats and
   say, seqno is 6
4. EP-engine resetted vbucket 228 in between
5. Index updater doesn't know about reset. It requested startseqno = 0,
   endseqno = 6

In order prevent stream from hanging, stream is requested with a special
flag that tells server to set endseqno. This flag needs to be set only
when we are requesting stream for a partition first time (ie, paritition
version is [{0, 0}]).

Change-Id: I292ed61fdbfc37b3c89edf32ec22527414e36187