Release fds properly for views with reduce queries
View group file descriptors are reference counted by the number of callers
and interested callers acquire an fd and increments the refcount.
For reduce enabled views, fd reference count is not decremented correctly
after the application discards that fd. This causes fd leaks in case of
keep-alive connections used for queries.

Change-Id: I20a74202587198da46b81e2e722dcbf9821d7b08