MB-12110 Extract snapshot marker type correctly
Snapshot type is determined by extracting the corresponding bits
in the marker type bit fields from snapshot marker upr message.

The current implementation checked for equality of snapshot marker
type field with on-disk or in-memory marker type. But, EP-Engine
may send snapshot marker with other bits set as well. In that case
snapshot marker is ignored and snapshot boundaries are missed out
by the updater. Since the snapshot merger fails to detect the
snapshot boundary, multiple entries for a document ends up in
the updater pipeline and duplicates are added into the index.

Change-Id: Ic58df6ec438dc3b2791fe4225a95748737c781e2