Use round-robin scheduling for mapreduce nif workers
Using random work allocation to mapreduce nif worker threads may
cause contention on worker threads and work distribution may not be
uniform. Considering the workload, when a map context is created
the updater uses that context to perform a number of map operations
during the lifetime of the updater. Hence, the best work distribution
may be to assign one worker thread to a map context and use the
same worker for all map tasks for that context. The worker thread
can be assigned in round-robin fashion to each map context created
by the updater.

Change-Id: I70d4c186086c42ad4f06e00a243a08e8757bd94e